refactor(esci2): extract withEsci2UnlockOnDestroy to transport.ts
Pulls the protocol-aware TLS-socket wrapper out of the orchestration
shell into a sibling adapter module. Scanner shell drops from 217 LOC
to 117. The wrapper is renamed from `withUnlockOnDestroy` to
`withEsci2UnlockOnDestroy` to make its protocol-specificity explicit
— ESC/I will need its own adapter (different close-timing rules), not
a shared one.

Wrapper unit tests move alongside as `transport.test.ts`. Scanner
test file keeps the integration-style fingerprint-pin and async-
fatal-event coverage.

Spec §3.1 and plan now record the M2 architectural finding: the
original "no wrappers needed / ~30 LOC shell" line was overoptimistic
because real Epson panel hygiene depends on application-level
teardown timing. Each protocol gets its own transport adapter; future
ESC/I-2-over-plain-TCP factories reuse this same one. M3 should
mirror the pattern with src/esci/transport.ts.
