+++ docs/superpowers/plans/2026-05-06-v0.4.0-architecture.md
@@ -4,7 +4,7 @@
 
 **Goal:** Extract a shared, transport-blind scan-session engine; refactor the ESC/I-2 scanner onto it; rewrite the ESC/I scanner onto it. Eliminates ~80 LOC of async-encode glue, the `getState()` TypeScript-narrowing workaround, five copies of the `ESC e + param + double-ACK` pattern, and the `recvBuffer` strategy divergence between the two scanners. Establishes a single `Graph<Ctx>` data structure + `SessionTransport` interface that lets a future ET-2750 (issue [#43](https://github.com/mtheuma/epson2paperless/issues/43)) land as a one-line factory swap.
 
-**Architecture:** A new `src/scan-session.ts` module owns generic graph execution, IS-packet framing, recv-buffer accumulation, rolling-timeout management, page-flush barriers, and EXIF orientation injection. Each scanner becomes a thin (~30 LOC) orchestration shell that builds a transport factory, calls `runScanSession({ graph, initialCtx, transportFactory, … })`, and translates the engine's `Result` to the existing promise-rejection contract. Per-protocol logic lives in `src/esci2/graph.ts` and `src/esci/graph.ts` (graph builders + protocol-specific helpers).
+**Architecture:** A new `src/scan-session.ts` module owns generic graph execution, IS-packet framing, recv-buffer accumulation, rolling-timeout management, page-flush barriers, and EXIF orientation injection. Each scanner is an orchestration shell that builds a transport factory, calls `runScanSession({ graph, initialCtx, transportFactory, … })`, and translates the engine's `Result` to the existing promise-rejection contract. Per-protocol logic lives in `src/esci2/graph.ts` (and `src/esci/graph.ts` after M3) — graph builders plus protocol-specific helpers. Protocol-aware transport behavior (e.g. ESC/I-2's UNLOCK-on-destroy + post-FIN error suppression) lives in a sibling adapter module (`src/esci2/transport.ts` exporting `withEsci2UnlockOnDestroy`); the shell wires raw socket → adapter → engine. **M2 actual scanner-shell size is ~120 LOC** (the original ~30 LOC target predated the transport-adapter finding — see spec §3.1 M2 finding); M3 should follow the same pattern with a sibling `src/esci/transport.ts`, not unify with ESC/I-2's adapter (close-timing rules differ between the two protocols).
 
 **Tech Stack:** Node.js 24+ / TypeScript 6 / Vitest 4 / Zod 4 / sharp 0.34 / pdf-lib 1.17. No new runtime dependencies.
