+++ docs/superpowers/specs/2026-05-06-v0.4.0-architecture-design.md
@@ -60,12 +60,14 @@ export interface SessionTransport {
 }
 ```
 
-Both Node `tls.TLSSocket` and `net.Socket` already satisfy this structurally — no wrappers needed. Tests use a tiny fake transport (no real stream semantics required).
+Both Node `tls.TLSSocket` and `net.Socket` satisfy this interface structurally — the engine never reaches into either. Tests use a tiny fake transport (no real stream semantics required).
 
 **Engine responsibility:** "Given an already-connected byte stream, run this graph."
 **Factory responsibility:** "Connect however this protocol requires, then return a stream."
 
-The dispatcher (`src/protocol-probe.ts` + `src/startup.ts:dispatchScanSession`) keeps probe / auto-detection logic. The factory is dumb. ET-2750 lands as a third dispatcher case (ESC/I-2 graph + plain-TCP factory) without engine or factory primitive changes.
+> **M2 finding (2026-05-07).** The original "no wrappers needed" line was overoptimistic. ESC/I-2 panel hygiene depends on application-level teardown timing — sending UNLOCK before half-close, swallowing benign post-FIN ECONNRESET / EPIPE, no-op'ing destroy after a graceful end so close_notify lands. None of that belongs in the generic engine; none of it belongs in the orchestration shell either. Each protocol that needs this gets its own **transport adapter** module (`src/esci2/transport.ts` exporting `withEsci2UnlockOnDestroy(socket: tls.TLSSocket): SessionTransport`). The factory wires raw socket → adapter → engine. Future ESC/I-2-over-plain-TCP gets a sibling factory in `scanner.ts` reusing the same adapter unchanged. ESC/I, when ported to the engine in M3, gets its own adapter (`src/esci/transport.ts`) — do not unify with ESC/I-2's, the close-timing rules differ.
+
+The dispatcher (`src/protocol-probe.ts` + `src/startup.ts:dispatchScanSession`) keeps probe / auto-detection logic. The factory is small. ET-2750 lands as a third dispatcher case (ESC/I-2 graph + plain-TCP factory + same `withEsci2UnlockOnDestroy` adapter) without engine or graph changes.
 
 ### 3.2. Graph data structure
 
@@ -392,20 +394,23 @@ src/
 
 ### 4.2. Scanner orchestration shells (M2 / M3)
 
-After M1–M3, each `scanner.ts` is a thin shell:
+After M1–M3, each `scanner.ts` is a thin shell — wires session config to the transport factory, threads the protocol-specific transport adapter (§3.1 finding), and translates the engine's `Result` to the existing promise-rejection contract:
 
 ```ts
-// src/esci2/scanner.ts (after M2; ~30 LOC)
+// src/esci2/scanner.ts (M2 actuals: ~120 LOC after extracting the
+// transport adapter to ./transport.ts; the original ~30 LOC target
+// assumed no protocol teardown logic — see §3.1 M2 finding)
 
 export async function runEsci2Scan(
   session: ScanSession,
   socketFactory: TlsSocketFactory = tls.connect,
 ): Promise<void> {
-  const transportFactory = makeTlsTransportFactory(session, socketFactory);
+  // Validate temp-dir base (fail-fast before TLS handshake).
+  // ... fs.accessSync(tempBase) ...
   const result = await runScanSession({
     graph: esci2Graph,
     initialCtx: { duplex: session.duplex /* ... */ },
-    transportFactory,
+    transportFactory: makeTlsTransportFactory(session, socketFactory), // wires withEsci2UnlockOnDestroy
     outputDir: session.outputDir,
     tempDir: session.tempDir,
     sessionTs: resolveSessionTimestamp(new Date(), session.outputDir),
@@ -416,7 +421,7 @@ export async function runEsci2Scan(
 }
 ```
 
-The graph + helpers + Ctx live in `./graph.ts`. The orchestration shell only wires the transport factory and translates the engine's `Result` to the existing promise-rejection contract.
+The graph + helpers + Ctx live in `./graph.ts`. The protocol-aware transport adapter lives in `./transport.ts`. The shell composes them.
 
 ## 5. Issue #43 fold-in
 
@@ -425,8 +430,8 @@ The transport-orthogonality constraint is satisfied by the design above; this se
 ET-2750 (hypothesis 1: ESC/I-2 over plain TCP) lands as:
 
 1. A new dispatcher branch in `src/protocol-probe.ts` / `src/startup.ts:dispatchScanSession` recognising "ESC/I-2 protocol script + plain TCP transport." Detection rule TBD by the v0.3.1 protocol-diagnose mode output.
-2. `src/esci2/scanner.ts` exposes a second factory variant: `makeTcpTransportFactory(session)` alongside the existing `makeTlsTransportFactory`. Cert-fingerprint logic stays in the TLS variant only.
-3. Same `esci2Graph` runs over the new factory. **No engine changes. No graph changes.**
+2. `src/esci2/scanner.ts` exposes a second factory variant: `makeTcpTransportFactory(session)` alongside the existing `makeTlsTransportFactory`. Cert-fingerprint logic stays in the TLS variant only. Both factories wrap their raw socket in the existing `withEsci2UnlockOnDestroy` adapter from `src/esci2/transport.ts` — same panel-hygiene rules apply over plain TCP.
+3. Same `esci2Graph` runs over the new factory. **No engine changes. No graph changes. No transport-adapter changes.**
 
 Hypothesis 2 (eSCL on port 80) would NOT use this engine — eSCL is HTTP/SOAP and shares no IS-framing or ESC/I-2 vocabulary with our engine. It would be a parallel implementation under `src/escl/` sharing only the output-tail boundary. Group A doesn't anticipate this; it would be a separate spec.
